+++ playwright-tests/createNft/actions.ts
@@ -2,16 +2,16 @@ import { Page } from '@playwright/test';
 import { DataTestIdsEnum } from '../../src/localConstants/dataTestIds.enum';
 import { GlobalDataEnum } from '../utils/enums';
 
-interface NftCreation {
+type NftCreationType = {
   page: Page;
   colectionName?: string;
   nftName?: string;
-}
+};
 export const createNft = async ({
   page,
   colectionName = 'col',
   nftName = 'TestNft'
-}: NftCreation) => {
+}: NftCreationType) => {
   await page.waitForTimeout(5000);
   await page.locator('.css-19bb58m').click();
   await page.getByRole('option', { name: colectionName }).first().click();