+++ src/config/sharedNetworks.ts
@@ -11,8 +11,7 @@ export const sharedNetorks: NetworkType[] = [
     extrasApi: 'https://devnet-extras-api.multiversx.com',
     sampleAuthenticatedDomains: ['https://devnet-api.multiversx.com'],
     sovereignContractAddress: '',
-    walletAddress: 'https://devnet-wallet.multiversx.com',
-    WEGLDid: ''
+    walletAddress: 'https://devnet-wallet.multiversx.com'
   },
   {
     default: false,
@@ -23,8 +22,7 @@ export const sharedNetorks: NetworkType[] = [
     extrasApi: 'https://extras-api.multiversx.com',
     sampleAuthenticatedDomains: ['https://api.multiversx.com'],
     sovereignContractAddress: '',
-    walletAddress: 'https://wallet.multiversx.com',
-    WEGLDid: ''
+    walletAddress: 'https://wallet.multiversx.com'
   },
   {
     default: false,
@@ -35,20 +33,20 @@ export const sharedNetorks: NetworkType[] = [
     extrasApi: 'https://testnet-extras-api.multiversx.com',
     sampleAuthenticatedDomains: ['https://testnet-api.multiversx.com'],
     sovereignContractAddress: '',
-    walletAddress: 'https://testnet-wallet.multiversx.com',
-    WEGLDid: ''
+    walletAddress: 'https://testnet-wallet.multiversx.com'
   },
   {
     default: false,
     id: 'sovereign',
     name: 'Sovereign',
-    apiAddress: 'https://api-sovereign-test.elrond.ro',
+    apiAddress: 'https://localhost:3002',
     gatewayUrl: '',
-    extrasApi: 'https://extras-api-sovereign-test.elrond.ro',
-    sampleAuthenticatedDomains: ['https://api-sovereign-test.elrond.ro'],
+    extrasApi: 'https://localhost:3001',
+    sampleAuthenticatedDomains: ['https://localhost:3002'],
     sovereignContractAddress: '',
-    walletAddress: 'https://wallet.voyager1.dev',
+    walletAddress: 'https://localhost:3000',
     WEGLDid: 'WEGLD-bd4d79',
+    hrp: 'erd',
     faucet: true,
     hasRegisterToken: true,
     hasSovereignTransfer: true