+++ src/redux/slices/network.ts
@@ -2,7 +2,7 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { networks } from 'config';
 
 export interface NetworkType {
-  WEGLDid: string;
+  WEGLDid?: string;
   apiAddress: string;
   default: boolean;
   extrasApi: string;